benchmarks: add comparisons to confluent-kafka-go & sarama
I was not expecting confluent-kafka-go to be so slow, but I guess the
cgo and then channel overhead quickly adds up.

For sarama, I was not expecting consuming to be so fast. But, franz-go
still beats it by 1.5x.

This also drops down the fetch max bytes in the franz-go benchmark,
because decompressing inflated responses a lot, which ate up a
significant amount of memory.