metadata: limit retries, bump produce load errors on failure
Previously, if we could not dial a broker, then we would *not* bump load
errors on buffered records. This meant that we would wait forever, even
if produce retries was limited.

Now, we limit retries for internally-caused metadata requests to 3, and
we bump errors on all partitions and any waiting unknown topic on
metadat failure.

The number 3 keeps the internal retries tolerant to single-broker
problems (e.g., cannot dial seed 1), but also limits things to be small
enough to not retry for an unreasonable amount of time.

This changes two logging messages to better reflect the ambiguity of
what is going on.