client: remove BrokerConnDeadRetries option; drop retries to 20
A recent commit dropped the number of retries by default from unlimited
to 30. Originally, produce retries was not separate from retries, so
retries was required to be unlimited.

Now that we have dropped the default retries to a reasonable number,
there is less of a need for BrokerConnDeadRetries, and the option is
just confusing: 3 retry options?

Removal of this option simplifies shouldRetry and allows removing
errDeadConn. This will not affect isRetriableBrokerErr because we unwrap
into the underlying os.SyscallError anyway, and we retry then.