client: large breaking change for consuming APIs
This change is a long time coming, and regrettably, it is the most
difficult part of this release.

This removes Assign*, in favor of just receiving an initialized client
that is ready to consume. All consuming options have been moved to be
options to pass to NewClient.

This removes one weird area of the client, where you had to initialize a
client and then assign what to consume. It would be better to just
initialize into consuming up front.

This does remove the ability to reassign partitions at runtime. That is
an at-best niche use case, and the existence of it historically led to
bugs. It keeps the API and the client simpler to just have a static
input set.

Now, to consume from a group, you will need the ConsumerGroup option.
We also remove the Offset argument from ConsumeTopics, instead relying
solely on the ConsumeResetOffset option now.

Additionally, we now pass the client to the group function callbacks,
because now that they are top level options, users do not have a chance
to know what the client is in the callback unless the client gives it to
the user.

Overall, this change may seem complex, but I hope that it is a net win
in the end.