(minor) breaking api change: BlockingCommitOffsets => CommitOffsetsSync
After the prior comits adding CommitRecords and
CommitUncommittedOffsets, there is a much more valid reason for this
change: consistency & docs.

ProduceSync was added a bit ago, and I've been well aware of the
inconsistency between BlockingCommitOffsets and ProduceSync. Ideally,
they would have the same type of name. I think I initially started with
CommitOffsetsSync, but thought Blocking was better. I'm not sure that is
the case anymore.

Second, now that we have four (4) ways to commit, I think it would be
very good if all four methods are next to each other in documentation.
Although only two methods are recommended (CommitRecords and, better,
CommitUncommittedOffsets), it is great now to have all methods start
with Commit for documentation proximity (since we cannot control godoc's
alphabetical ordering).

Lastly, a side benefit of breaking this is that now anybody using it
will be forced to notice the change and then potentially see that an
even better API was added: CommitUncommittedOffsets.