client: make public ErrClientClosed
I noticed that there is no obvious way to detect the client closing when
polling fetches / records. The next commit will introduce injecting a
fake error on poll when the client is closing, but in order to allow a
poll loop to detect client closing, we need to make the error public.