add pluggable kgmetrics, kprom, kzap packages & examples
This creates three new packages in the new plugin directory:
- kgmetrics
- kprom
- kzap

These packeages are each modules in their own right so that a general
user of franz-go does not need to update their dependencies to include
unrelated packages.

The metrics packages tries to take a relatively sane yet thorough
approach to metrics, and the usage attempts to be simple.

This then adds three runnable examples demonstrating how to use the
plugins, all of which I've tested locally myself.

Finally, this updates some wording in READMEs, and renames & expands
(with a produce flag) the existing prometheus example.