producer: allow configurable max inflight when disabling idempotency
This allows for higher throughput even if you want to disable
idempotency, at the increased risk of more duplicates and potentially
out of order records.

This is a reintroduction of commit
229f9da2906dce9813c4ad9d93385712fd95bd01,
which was reverted for mostly being useless in
d1ba10360f457b4007cb2910bc23fc37a07b05cd
and was re-requested in #129.

The API is so minor and it is easy enough to document it shouldn't be
used really, so may as well add it back. However, we will be going with
a longer and more explicit name for the option, which will make it
clearer in code if anybody uses the option, and will help disincentize
usage.