group balancing: address KAFKA-12898
We now just wholesale ensure everything is sorted when balancing. This
will marginally slow things down.

The main problem is that Kafka considers rebalancing whenever a group
join changes. Really, we should only be rejoining when we are
rebalancing anyway, but ensuring deterministic output helps reduce a
vector for an accidental rebalance, and will make the logs nicer.