client: log "seed #" rather than a large negative for seed brokers
Previously, seed brokers logged with highly negative numbers, which was
confusing. Now, we log "seed 0", "seed 1", etc.

We still use negatives in int32 returns, but that is well documented:
basically, BrokerMetadata.NodeID documents the negative usage for seeds.

Lastly this cleans up some wording on producing to an unknown topic for
the first time.