sink: drop the drain loop sleep from 5ms to 50microsec
The point of this sleep is to ensure that two records produced back to
back in code do not get produced in separate batches. 5ms is overkill,
50microsec is hopefully enough to just reschedule the goroutines and
allow the two records to produce into the same batch. Overall this only
impacts the first two sequential records, under high volume the loop
will not stop.