broker: use ErrClientClosed more properly
Some internal requests use the client context to issue requests, so
closing the client could select the request context done case. This
would show errors saying "context canceled" instead of the more
appropriate "client closed".

Now, if the request context is equal to the client context, then we know
this is an internal request and the client is closed, so we return the
more appropriate error.