support KIP-516 for Fetch
This is fairly simple for the fetch request if we do not consider a
topic being deleted and recreated. If we do, then things get complicated
in the metadata code.

I'm not sure we want to support from a client a topic being deleted and
recreated, so we will punt on that until more of KIP-516 comes into
play.