hooks: add HookNewClient
This new hook can notify hook implementations _of_ the new client.
Previously, a hook had to be declared before the client, and it was
difficult for a hook to know of the client. Now, a hook can hook into
extra behavior because of the new client: specifically, a hook can be
notified of the client and install two extra metrics for the buffered
produce and fetch records.