producing: only opt in to more than 1req if idempotent
Previously, we relied on the produce response being v4 or higher to say
"ok we are talking to kafka v1+ and we can always be idempotent!".

We actually only want to opt in to more requests only if we ourselves
are idempotent.

To detect this, we only opt up if the request was issued with a producer
ID AND the response is v4 or higher. Ensuring we issued with a producer
ID ensures that we are idempotent.

This also simplifies some logic around where the idempotent fields are /
how they are used.