kmsg: copy kbin into an internal package
To v1 the franz-go client, we need to have a dedicated module for kmsg
so that we can avoid stabilizing kmsg.

To do this, we need to not depend on kbin.

This duplicates kbin into an internal package. The next step is to
update the interface.go file and the generated file.