kmsg: un-depend on kbin package
This changes the SkipTags/ReadTags functions to take a new tiny
interface. Since we are not stabilizing kmsg, we do have a bit of
freedom in the future to potentially change these APIs (maybe to be
slices), but ideally that will not happen. It is just a bit awkard.

We also have internal{ReadTags,SkipTags} that operate on a non-interface
vendored kbin, which allows for better escape analysis and speed in our
actual protocol deserialization.