producer/txn: centralize sequence number resetting
If a producer ID failed, we needed to reset all sequence numbers.

For idempotent producers, this was done by just bumping the producer
epoch locally and resetting.

For transactional producers, this was done when ending a transaction,
even though transactional producers go through the same function to
reset a producer ID as the idempotent producer (the transactional
producer just goes through the request issuing branch).

Centralizing all isntances of resetAllProducerSequences makes things a
bit clearer and may prevent some inadvertent bug in the future.