plugins: depend on the latest franz-go (v0.9.0)
This is a proof of concept to show that depending on the latest franz-go
has no changes from the kmsg separation; the only change will be if you
use kmsg directly.

This has no practical change for the plugins, so I will not be tagging
these. The examples (and plugins) will be updated when I v1.0 the
library.