all: work around brokers that have inconsistent request versions
In some places, the client uses responses from some requests and assumes
that certain features are working. In particular, if the broker returns
a non-negative leader epoch in a metadata response, the client assumes
it can use that leader epoch to validate if a partition was truncated if
it moved to a different broker. That's not always the case, and the
client would permanently loop on failing epoch loading requests.

Now, we only opt in to KIP-320 in certain scenarios if we know that any
broker supports it. This could be finer grained, in that if one broker
in a cluster does not support and others do, we may have the same
failure semantics as before. But, that should ideally only be during
rolling upgrades.

This is basically a patch that should never trigger to work around what
is ideally a temporary bug in brokers.