client improvement: retry more when input brokers are invalid
If a user passed two seeds and one is invalid, we can still operate with
some degree of success. Previously, we would fail a request immediately
if it was attempted on the invalid seed. However, if the next seed is
valid, we should be able to recover, eventually load metadata, and
always be successful.

We will now retry on broker-specific network errors if they are not
context related, and if the next broker we would retry on is different.