sink: general improvements
This reorganizes some code to hopefully be a bit easier to understand.

* creates removeFromTxn, replaces the duplicate areas where a partition
is removed from a transaction

* removes requeueUnattemptedReq in favor of just using
handleRetryBatches. Unattempted requests may as well just be handled
similarly -- the only downside is a small backoff, which is likely
necessary anyway because unattempted requests are usually connection
failures, and we do not want to stampede reconnect.

* extracts the noack handling to a separate function. This makes reading
handleReqResp a bit simpler.

* extracts hook metrics to a separate type / function, which again makes
handleReqResp a bit simpler.

* removes tryResetFailingBatchesWith, moves the logic inline into
handleRetryBatches. tryResetFailingBatchesWith was used twice, once by
the now removed requeueUnattemptedReq. It is simpler to reason about
inline, now.

* panics if our inflight on the recBuf goes negative. This is just a
sanity check; it never should.