feature: add LeastLoadPartitioner
This was to test a theory of mine that production could continue if a
single broker was sniped without the client backing up entirely before
a metadata update reflected the down broker.

This is somewhat useful. Not strongly recommended due to load imbalance,
but this can be useful in some scenarios. Additionally, this adds a new
optional interface that people can use to implement more interesting
algorithms (at the expense of speed).