sink: simplify decInflight for batches in req
This simplifies the logic: grab all batches before issuing a request,
finish everything once handling the response is done. Importantly, this
also ensures that we do not double finish if we had a txnal req that
removed a partition from the request (because it would not be removed
from batchesFlat).