add RoundRobinPartitioner
This actually showed that OnNewBatch should not be a required method on
the Partitioner interface but instead should be an optional extension.
If it is required, then every message published to round robin goes to a
new batch:
- partition on 1?
  - no, new batch
- partition on 2
  - inc to 3
All partitioning starts on odd partitions, then completes on even.
Removing OnNewBatch gets us to true round robin.

Closes #66.