bump min go version to go1.15
The DialTLSConfig option bumped the min version to go1.15 from the
previous go1.13. The recent introduction of net.ErrClosed bumped the min
version to go1.16.

We can still support go1.15 by having a separate build file.

We will drop 1.13: it's very old, go1.17 is out now (so go1.15 is
technically unsupported), and some of the deps use go1.15 (notably,
compress).

Closes #69