consumer: make some fields visible to encoding
Out of the currently 114 logs in kgo that contain arguments, two of
those logs were not visible to anything using json encoding (and by
proxy, anything encoding through visibility of reflect): an Offset, and
listOrEpochLoads.

We add MarshalJSON and export two fields (in a private struct) where
necessary so that everything is visible with minimal modifications. We
also add String so that these structs are more nicely formatted in the
basic logger.