bugfix source: advance offsets even if we return no records
Commit 058f692 changed things to not buffer a fetch if the fetch
contained no records. Unfortunately, if a fetch has a small MaxBytes and
can only consume one record, and that one record is a control batch,
then we would never advance offsets.

We now advance offsets internally if we process a fetch successfully,
even if we do not buffer a fetch to be returned.