consumer group: document actual behavior for on revoked / lost
The code explicitly set a no-op OnLost internally if there was no user
OnLost specified. We now formalize this in documentation and in the
proper section in code. Unlike the Java client, we are not starting from
a history of always calling OnRevoked, so we can explicitly document
that OnLost must be specified if users are interested.

We now also ensure we duplicate the map passed to the callbacks so that
users can own them and mutate them as they desire. We also document that
the callbacks are never called concurrently.