kmsg: add EnumStrings(), ParseEnum() functions
This should make it easier to list valid options for the string value of
an enum, as well as easier to parse a string into an enum.

The parsing is aggressive in what we accept by lowercasing and stripping
all dots and underscores. I expect most input to just match based off
lowercasing, but a stray _ shouldn't affect anything.