producer: fail topics that repeatedly fail to load after 5 tries
If a record is produced to a topic that cannot load, it makes no sense
to continue trying to load it. Rather than trying indefinitely up to
the RecordDeliveryTimeout or the RecordRetries limit, we now fail after
5 load attempts. Given that we round-robin the brokers we load metadata
from, and given that this unknown load only happens on the first record
for a topic, 5 tries should be a safe default for any produce.