message set v1: bit 4 is timestamp type, not bit 3
As a result of misunderstanding the original Kafka protocol, and never
having any user that used LogAppendTime for v1 message sets, and message
sets only using the first two bits for compression, I mistakenly thought
the third bit was the timestamp bit.

Turns out it's the fourth bit. Kafka just never used the third bit,
although it was reserved.

This cleans up some wording in the message set / record batch
documentation to make which bits are used clearer, no longer mangles the
timestamp bit for v1 message sets, and allows the last four bits to be
present for v1 (rather than the last three) and the last three bits for
v0 (rather than the last two).