kgo: add RecordFormatter, RecordReader
These are niche functions, but are quite powerful in the event they are
needed. The current specification or reading / headers could be a little
bit better, but as is, the functions are useful.

This is basically a straight rip from the kcl repo, with modifications
to make things more consistent / have less shortcuts.

This also adds tests for most of the input/output to ensure things
actually work.