config: default to kgo dep version in ApiVersions
We can use debug.BuildInfo to determine which kgo version the user is
using.

We differ from sarama here by using the kgo version, rather than the
main module's version. This is more accurate to the client software name
/ version. Sarama uses the name "sarama", which will make it difficult
to match the version to the running program, since the version will be
for the binary, not the sarama library.