kerr: make BrokerNotAvailable retriable
This is not retriable per Kafka itself, but both the Java client and the
librdkafka client handle this as retriable. In either of those clients,
this error will force a metadata refresh (in Java, it looks like this
error is only handled in the context of metadata in the first place).

In this client, we will just accept the error as retriable outright.
Looking at Kafka proper, there's not many instances where this is
returned anymore, potentially only for the raft client?

Making this retriable allows InitProducerID to fail gracefully (and thus
internally retry). Redpanda currently returns broker_not_available in
some cases here.