errors: make errProducerIDLoadFail retriable
When a producer id load fails with a retriable error, we swallow the
error and instead return errProducerIDLoadFail. This gets used when
bumping repeated load failures on partitions, and since this internal
error does not look retriable, we fail records.

We now make this a retriable broker error (since it is), which allows
the producer ID load to fail and be retried.