kgo: only run req/resp handling goroutines when needed
Previously, kgo would start a handleReqs goroutine whenever an internal
newBroker was created. Any time handling reqs opened a connection, a
handleResps goroutine was begun.

Per broker, this meant usually two goroutines permanently hanging around
(one req, one either produce or consume), and occasionally a third (for
metadata/admin).

I noticed two goroutines that were hanging on handleReqs for 28K minutes.
This was the seed broker: it handled a request at the start and then was
never used again.

The changes here now only start goroutine loops when necessary, and stop
them when no work remains. The new "ring" types are basically a blend of
the previous channel and of the seqReqResps list. This is in fact a
better optimization for seqReqResps, so we are able to use this same
ring there and remove it constantly reallocating small arrays.

This also simplifies the connection dying aspect of brokers, and (in
broker code, not in the ring code) simplifies the locking considerably.
Previously, we basically had two locks: one on the dead field, and one
internally to the channel. We now combine those into one.