kadm: use more named returns, add more helpers
Some of the existing slice returns are better suited as maps, so we
switch some lists to maps. The configs bits remains as slices because
configs can be nameless (cluster-wide brokers).

We also now use named types for a lot more returns, and on these named
types we provide the standard Sorted helper function as well as a new On
helper function. The On helper is more useful when you do an operation
on one item.