client: lower some default timeouts
* ConnTimeoutOverhead has been lowered from 20s to 10s. This will allow
quicker detection of hung writes. For most purposes, this will be a fine
timeout, and requests will retry if necessary. For extremely bad
connections, users may need to raise this.

* RetryTimeout has been lowered from a default of 1m to 30s, with
JoinGroup, SyncGroup, and HeartbeatRequest being exempted to the session
interval by default. For requests, this will allow quicker detection of
hung brokers and will not retry as long. Internally in the client, this
*should* result in just extra logging in most cases. For groups, this
will avoid trying to stay in a group longer than the group would be
valid.

For #92.