broker: add two new connection types, cxnGroup and cxnSlow
As a group consumer, we expect 1 active join group request, and then 1
active sync group request. Both of these requests can hang. If a client
wants to update metadata while either of these requests are hanging,
and the metadata happens to choose the same connection, then the
metadata request will be blocked. We do not want this.

If an admin issues a request that contains a TimeoutMillis field, it is
possible that the request will take a while to execute. We do not want
to be blocked if this is so.

To fix both of these scenarios, we add two new connection types:
cxnGroup, which specifically handles Join and Sync, and cxnSlow, which
handles anything with timeout millis.

For #92.