consuming, group: improve logging, simplify code
The client long ago logged why assignments happened. That was removed
for being verbose, but as it turns out, the why is very valuable.

This re-adds the why. We avoid doubly logging when transactions finish
by avoiding logging if the reason is empty.

The Offset type now has a String method to more tersely represent what
the offset is.

We also move the consumer.go unset code directly into LeaveGroup, the
only spot it was used, and remove some unneeded code.

As well, we simplify some commit logic: because g.commit is meant to be
async, we do not need to `go` it.