add more context to metadata reloads on inner partition errors
Two cases can trigger loads of metadata due to inner errors, of which
there can be many. We added the first error's reason for fetching, but
we can do better and add every error.

This gives us larger messages that now print every error. We collapse
errors as much as possible, which should be the usual case for
kerr.Error.