client: support non-zero/one node IDs
The client previously assumed that broker node IDs started from zero.
This also inadvertently supported 1, because 0++ is one, and we would
check that often.

If broker IDs start at 2+, then 0 and 1 lookups will always fail, and we
fallback to seeds. If only one seed is specified, we essentially pin all
requests to one broker. This is a bug.

This commit splits seed brokers from non-seed brokers. We always store
the nodes in sorted order, so when we need to lookup a node, we can
sort.Search, which will keep lookups fast.

Saving brokers a bit more complicated.