source: use forgotten topics for sessions
Previously, if we paused a partition, we would still receive data for
that partition in fetch responses because we did not signal that we are
no longer consuming the topic. The would result in wasted skipped data.

As well, if a partition moved from one broker to another, we never
"forgot" it on the old, and hygiene wise, that meant we would still be
asking for it.

We now use forgotten topics properly so that if we used something
previously in the session but are not now, we forget it and properly
track our state.

In a way, this completes KIP-227, whereas before it was mostly done.