kadm: return OffsetResponses from FetchOffsetsForTopics
This allows the return to be directly used in CalculateGroupLag. The
purpose of the function was for consumers that manage assigning
partitions outside of Kafka, so they could FetchOffsetsForTopics for
topics that were not yet committed. This same functionality may be
useful for lag, which also uses the OffsetResponse.Error field.

By returning OffsetResponses, the user will have to do Into twice, but
that is easy (perhaps OffsetResponses just needs a direct IntoKgo).