kmsg: reuse prior slices in StickyMemberMetadata.ReadFrom
We can drop allocation count a good bit in the sticky package by having
better slice reuse here in ReadFrom.

There is not much gain on other types because I do not expect other
types to reuse the input so much. But, we know we do on sticky member
metadata, so...